update the structure
